refactor(plugin): rename VODProvider to verseOfDayApi (C6)
The module holds Verse-of-the-Day fetch functions (getVod/getEnhancedVod),
not a BaseBibleAPIProvider subclass - the ...Provider name and provider/
placement implied an adapter it never was. Rename the file; update the
3 import sites. 200 tests pass.
