+++ packages/obsidian-bible-reference/src/main.ts
@@ -11,7 +11,7 @@ import { VerseLookupSuggestModal } from './suggesetor/VerseLookupSuggestModal'
 import { TranslationSwitcherModal } from './suggesetor/TranslationSwitcherModal'
 import { VerseOfDayEditorSuggester } from './suggesetor/VerseOfDayEditorSuggester'
 import { VerseOfDayModal } from './suggesetor/VerseOfDayModal'
-import { getEnhancedVod } from './provider/VODProvider'
+import { getEnhancedVod } from './provider/verseOfDayApi'
 import { splitBibleReference } from './utils/splitBibleReference'
 import { VerseOfDaySuggesting } from './verse/VerseOfDaySuggesting'
 import { getBibleVersion } from './data/BibleVersionCollection'
